refactor(85381): Move GestaoDeUsuario.service de volta para services globais
Esses serviços serão usados tanto na lista como no form de usuários pelo
menos. Por isso devem estar nos serviços compartilhados.
